refactor: extract Shell socket/session lifecycle into useShellSession hook (#2532)
Move the Shell page's ~750 lines of xterm setup, the Socket.IO shell:* protocol,
and the single-subscriber PTY attach/detach contract (pendingAttachRef
{target,generation} guards, claim vs user-intent, userIdleRef passive-idle
gating, mountedRef deferred-work guards, URL sync) into
client/src/hooks/useShellSession.js, which documents the full attach/detach
contract at the top of the file. Shell.jsx drops to a thin presentational shell
plus two extracted panels (TerminalHotKeys, ShellSessionTabs), following the
WritersRoom pattern. Behavior identical; generation guards preserved.
